Moves config command to CommandLine.Xplat and implements suggestions from PR
