+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Commands/CommandParsers.cs
@@ -21,7 +21,6 @@ namespace NuGet.CommandLine.XPlat
             ListVerbParser.Register(app, getLogger);
             RemoveVerbParser.Register(app, getLogger);
             UpdateVerbParser.Register(app, getLogger);
-            ConfigVerbParser.Register(app, getLogger);
         }
     }
 }
