+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Commands/Commands.xml
@@ -110,11 +110,4 @@
       <SingleValueOption Name="configfile" Help="Option_ConfigFile" />
     </Noun>
   </Verb>
-  <Verb Name="config">
-    <Noun Name="paths" Version="3.1.200"
-       Description="lists all the paths of NuGet configuration files that will be applied." >
-      <Argument Name="working-directory" Help="ConfigPathsWorkingDirectoryDescription"/>
-      <Example Title="List configuration files that will be applied in the current directory:" Command="dotnet nuget config paths" />
-    </Noun>
-  </Verb>
 </Commands>
