+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Commands/ConfigCommands/ConfigArgs.cs
@@ -1,0 +1,10 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace NuGet.CommandLine.XPlat
+{
+    internal partial class ConfigPathsArgs
+    {
+        public string WorkingDirectory { get; set; }
+    }
+}
