+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Commands/Verbs.cs
@@ -475,39 +475,4 @@ namespace NuGet.CommandLine.XPlat
         }
     }
 
-    internal partial class ConfigVerbParser
-    {
-        internal static void Register(CommandLineApplication app,
-                                      Func<ILogger> getLogger)
-        {
-            app.Command("config", ConfigCmd =>
-            {
-                ConfigCmd.Command("paths", PathsCmd =>
-                {
-                    CommandArgument workingdirectory = PathsCmd.Argument(
-                        "working-directory", Strings.ConfigPathsWorkingDirectoryDescription);
-                    PathsCmd.HelpOption("-h|--help");
-                    PathsCmd.Description = Strings.ConfigPathsCommandDescription;
-                    PathsCmd.OnExecute(() =>
-                    {
-                        var args = new ConfigPathsArgs()
-                        {
-                            WorkingDirectory = workingdirectory.Value,
-                        };
-
-                        ConfigPathsRunner.Run(args, getLogger);
-                        return 0;
-                    });
-                });
-                ConfigCmd.HelpOption("-h|--help");
-                ConfigCmd.Description = Strings.Config_Description;
-                ConfigCmd.OnExecute(() =>
-                {
-                    app.ShowHelp("config");
-                    return 0;
-                });
-            });
-        }
-    }
-
 }
