+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Strings.resx
@@ -821,4 +821,8 @@ Non-HTTPS access will be removed in a future version. Consider migrating to 'HTT
   <data name="Config_Description" xml:space="preserve">
     <value>NuGet configuration CLI</value>
   </data>
+  <data name="Error_PathNotFound" xml:space="preserve">
+    <value>The specified path '{0}' does not exist.</value>
+    <comment>0 - path</comment>
+  </data>
 </root>
