+++ src/NuGet.Core/NuGet.Commands/CommandArgs/VerbArgs.cs
@@ -96,9 +96,4 @@ namespace NuGet.Commands
         public string Configfile { get; set; }
     }
 
-    public partial class ConfigPathsArgs
-    {
-        public string WorkingDirectory { get; set; }
-    }
-
 }
