+++ src/NuGet.Core/NuGet.Commands/NuGet.Commands.csproj
@@ -65,4 +65,8 @@
   <ItemGroup>
     <Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="ConfigCommands\" />
+  </ItemGroup>
 </Project>
