+++ src/NuGet.Core/NuGet.Commands/PublicAPI/net472/PublicAPI.Unshipped.txt
@@ -1,7 +1,1 @@
 #nullable enable
-NuGet.Commands.ConfigPathsArgs
-NuGet.Commands.ConfigPathsArgs.ConfigPathsArgs() -> void
-NuGet.Commands.ConfigPathsRunner
-~NuGet.Commands.ConfigPathsArgs.WorkingDirectory.get -> string
-~NuGet.Commands.ConfigPathsArgs.WorkingDirectory.set -> void
-~static NuGet.Commands.ConfigPathsRunner.Run(NuGet.Commands.ConfigPathsArgs args, System.Func<NuGet.Common.ILogger> getLogger) -> void
